requirements.txt
azure-identity==1.15.0
# Flask[async]==2.3.2
openai==1.6.1
python-dotenv==1.0.0
quart==0.19.4
uvicorn==0.24.0
aiohttp==3.9.2
gunicorn==20.1.0
